+++ src/main/java/io/brightskies/loyalty/points/repo/PointsRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.brightskies.loyalty.points.entity.Points;
 
-public interface PointsRepo extends JpaRepository<Points, Integer> {
+public interface PointsRepo extends JpaRepository<Points, Long> {
 
 }
