+++ src/main/java/io/brightskies/loyalty/refund/repo/RefundRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.brightskies.loyalty.refund.entity.Refund;
 
-public interface RefundRepo extends JpaRepository<Refund, Integer> {
+public interface RefundRepo extends JpaRepository<Refund, Long> {
 
 }
